+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log (e.GPT)
 
+## 0.4.0
+
+- `.system` file in `${HOME}/.egpt` can now contain a default system prompt, thanks to [jbonot](https://github.com/egomobile/e-gpt/issues/1)
+
+## 0.3.0
+
+- implement `sql` command
+
 ## 0.2.0
 
 - implement `environment` command